+++ CHANGES.md
@@ -1,5 +1,6 @@
 ### 1.0.1
 * Added AspectRatioContainer.
+* Added highlightActor to PopTable to draw widgets above the stageBackground drawable for demonstrations.
 
 ### 1.0.0
 * Moved FreeTypeSkin and FreeTypeSkinLoader to separate subproject.